+++ docs/mdfiles/changelog.md
@@ -2,6 +2,7 @@ Change Log
 ==========
 # 4.3
 * The type of `conferenceClient.publish` method's option is changed from `AudioEncodingParameters` and `VideoEncodingParameters` to `RTCRtpSendParameters`.
+* `audio` and `video` of `PublicationSettings` is changed from `AudioPublicationSettings` and `VideoPublicationSettings` to `AudioPublicationSettings` array and `VideoPublicationSettings` array.
 
 # 4.0.2
 * Fix a compatibility issue for Chrome 69 beta.